[202412] [PR:24053] Integrate SHOW API probe into telemetry watchdog (#1708)
<!--
Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

** Make sure all your commits include a signature generated with `git
commit -s` **

If this is a bug fix, make sure your description includes "fixes #xxxx",
or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
Manual cherrypick for
https://github.com/sonic-net/sonic-buildimage/pull/24053 due to
conflict.

Integrate SHOW API probe into telemetry watchdog. We want the watchdog
to catch the bad SHOW command implementation.

##### Work item tracking
- Microsoft ADO **(number only)**:34793098

#### How I did it
Add gnmi_get into telemetry-watchdog, and call gnmi_get in the watchdog
http service for each probe, as long as one xpath fails, the probe will
return 500.

The watchdog will read xpath list from /cmd_list.json and environment
variables.

#### How to verify it
Run the telemetry-watchdog container, call curl http://127.0.0.1:50080/
to check if the output is expected or not.

root@bjw-can-2700-2:/# curl -s http://127.0.0.1:50080/ |jq .
{
  "check_telemetry_port": "OK",
  "xpath_commands": [
    {
      "xpath": "clock",
      "success": true,
      "error": null,
      "duration_ms": 200
    }
  ]
}

<!--
If PR needs to be backported, then the PR must be tested against the
base branch and the earliest backport release branch and provide tested
image version on these two branches. For example, if the PR is requested
for master, 202211 and 202012, then the requester needs to provide test
results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111
- [ ] 202205
- [ ] 202211

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
Ensure to add label/tag for the feature raised. example - PR#2174 under
sonic-utilities repo. where, Generic Config and Update feature has been
labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on
https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)