[action] [PR:21907] sonic-yang-mgmt: uses clause with leaf-list, choice not honored (#1731)
#### Why I did it

When a uses clause imports a grouping, it was only processing leaf entries and ignoring leaf-list and choice clauses. That means that for instance in bgp route maps, route_map_in and route_map_out validations would fail.

Honoring the uses `refine` clause is now also honored which is depended upon in sonic-utilities.

This now precompiles the uses clause and integrates it as if the uses clause was not part of the schema as multiple end users were having to do this additional processing.

This fixes that behavior and adds test cases to ensure it doesn't regress in the future.

This is the proper fix, replacing #21078 that just worked around it.

Removal of `uses` logic in sonic-utilities here: https://github.com/sonic-net/sonic-utilities/pull/3814

Fixes https://github.com/sonic-net/sonic-buildimage/issues/22382

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

Added leaf-list lookup.

#### How to verify it

See test cases pass

#### Which release branch to backport (provide reason below if selected)

- [X] 202405
- [X] 202411

#### Tested branch (Please provide the tested image version)

master as of 20250304

#### Description for the changelog
sonic-yang-mgmt: uses clause with leaf-list, choice not honored

#### Link to config_db schema for YANG module changes
N/A

#### A picture of a cute animal (not mandatory but encouraged)
Signed-off-by: Brad House (@bradh352)