[202412][Broadcom] Upgrade xgs SAI version to 13.2.1.14 (#1772)
<!--
Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

** Make sure all your commits include a signature generated with `git
commit -s` **

If this is a bug fix, make sure your description includes "fixes #xxxx",
or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
This is a manual cherry-pick of
https://github.com/sonic-net/sonic-buildimage/pull/24397
Upgrade the xgs SAI version to 13.2.1.14 to include the following
changes:
- 13.2.1.14: Add Tam
##### Work item tracking
- Microsoft ADO **(number only)**: 35750547

#### How I did it
Upgrade the SAI version in the sai.mk file.
#### How to verify it

<!--
If PR needs to be backported, then the PR must be tested against the
base branch and the earliest backport release branch and provide tested
image version on these two branches. For example, if the PR is requested
for master, 202211 and 202012, then the requester needs to provide test
results on master and 202012.
-->



Signed-off-by: zitingguo-ms <zitingguo@microsoft.com>

Signed-off-by: zitingguo-ms <zitingguo@microsoft.com>